Moved WordCounter class into doceditor file
