+++ nw/gui/__init__.py
@@ -31,7 +31,6 @@ from nw.gui.elements.viewdetails import GuiDocViewDetails
 # Tools
 from nw.gui.tools.dochighlight import GuiDocHighlighter
 from nw.gui.tools.optionstate import OptionState
-from nw.gui.tools.wordcounter import WordCounter
 
 __all__ = [
     "GuiBuildNovel",
@@ -58,5 +57,4 @@ __all__ = [
     "GuiDocViewDetails",
     "GuiDocHighlighter",
     "OptionState",
-    "WordCounter",
 ]
