+++ nw/gui/tools/__init__.py
@@ -2,10 +2,8 @@
 
 from nw.gui.tools.dochighlight import GuiDocHighlighter
 from nw.gui.tools.optionstate import OptionState
-from nw.gui.tools.wordcounter import WordCounter
 
 __all__ = [
     "GuiDocHighlighter",
     "OptionState",
-    "WordCounter",
 ]
